Made lots of fixes to the module
I've made lots of fixes. As I do not have a deep knowledge of the
module, it is highly advisable if anyone (Skn3?) Could take a deeper
lool to my fixes.
I've marked them all with a Jungle  Ide note, so they're ultra easy to
locate in a glance.
all marked with 'note: Ziggy: are fix suggestions that should allow the
module to compile.
Then, there are much more crytical fixes marked with 'note: FIX TODO
that require some thought, as they seems like a design logical bugs,
such as Void methods that have Return statements with actual return
values!! or the like. Also, the creation of an abstract Map with a New
keyword, etc.
It requires additional attention by someone that has deeper knowledge on
this module internals.
Hope it's helpful!